C prototipo/Backend_API/requirements.txt
fastapi
uvicorn[standard]
paho-mqtt
mysql-connector-python
python-dotenv
sqlalchemy
pydantic-settings
pymysql
cryptography
python-dateutil>=2.8.0